add func for creating batches, rename date_utils module to utils
