+++ suncal/models/icalendar.py
@@ -7,7 +7,7 @@ from uuid import uuid4
 from pydantic import BaseModel  # pylint: disable=E0611
 from pydantic import validator
 
-from suncal.date_utils import aware_datetime_to_ical_date_with_utc_time
+from suncal.utils import aware_datetime_to_ical_date_with_utc_time
 from suncal.models.googlecal import GoogleCalEvent
