+++ tests/test_utils.py
@@ -1,8 +1,8 @@
 import datetime as dt
 
-from suncal.date_utils import aware_datetime_to_ical_date_with_utc_time
-from suncal.date_utils import create_timezone_aware_datetime
-from suncal.date_utils import date_range
+from suncal.utils import aware_datetime_to_ical_date_with_utc_time
+from suncal.utils import create_timezone_aware_datetime
+from suncal.utils import date_range
 
 
 def test_date_range():
